+++ crates/icydb/src/lib.rs
@@ -169,7 +169,7 @@ pub mod __macro {
         runtime_value_into, runtime_value_map_collection_to_value, runtime_value_to_value,
         runtime_value_vec_from_value,
     };
-    pub use icydb_core::value::{Value, ValueEnum};
+    pub use icydb_core::value::{InputValue, Value, ValueEnum};
 }
 
 // re-exports